Revert "refactor(FileValidationInterceptor): merge FileValidationInterceptor into DamUploadFileInterceptor"
This reverts commit e66199843f02d58f9bbe3188a8d05298622db47a.
