+++ packages/api/cms-api/src/dam/files/dam-upload-file.interceptor.ts
@@ -1,24 +1,17 @@
-import { HttpException, Inject, Injectable, mixin, NestInterceptor, Type } from "@nestjs/common";
-import { ExecutionContext } from "@nestjs/common/interfaces/features/execution-context.interface";
-import { CallHandler } from "@nestjs/common/interfaces/features/nest-interceptor.interface";
+import { Inject, Injectable, mixin, NestInterceptor, Type } from "@nestjs/common";
 import { FileInterceptor as NestFileInterceptor } from "@nestjs/platform-express";
 import { MulterOptions } from "@nestjs/platform-express/multer/interfaces/multer-options.interface";
 import fs from "fs";
 import * as mimedb from "mime-db";
 import * as multer from "multer";
 import os from "os";
-import { Observable, throwError } from "rxjs";
-import util from "util";
+import { Observable } from "rxjs";
 import { v4 as uuid } from "uuid";
 
 import { CometValidationException } from "../../common/errors/validation.exception";
 import { defaultDamAcceptedMimetypes } from "../common/mimeTypes/default-dam-accepted-mimetypes";
 import { DamConfig } from "../dam.config";
 import { DAM_CONFIG } from "../dam.constants";
-import { svgContainsJavaScript } from "./files.utils";
-
-const readFile = util.promisify(fs.readFile);
-const unlinkFile = util.promisify(fs.unlink);
 
 export function DamUploadFileInterceptor(fieldName: string): Type<NestInterceptor> {
     @Injectable()
@@ -82,44 +75,8 @@ export function DamUploadFileInterceptor(fieldName: string): Type<NestIntercepto
             this.fileInterceptor = new (NestFileInterceptor(fieldName, multerOptions))();
         }
 
-        async validateFileContent(file: {
-            mimetype: string;
-            destination?: string;
-            filename?: string;
-            path?: string;
-        }): Promise<Observable<unknown> | undefined> {
-            if (file && file.path && file.mimetype === "image/svg+xml") {
-                const fileContent = await readFile(file.path, { encoding: "utf-8" });
-
-                if (svgContainsJavaScript(fileContent)) {
-                    // https://github.com/expressjs/multer/blob/master/storage/disk.js#L54-L62
-                    const path = file.path;
-
-                    delete file.destination;
-                    delete file.filename;
-                    delete file.path;
-
-                    if (path) {
-                        await unlinkFile(path);
-                    }
-
-                    return throwError(() => new HttpException("Rejected File Upload: SVG must not contain JavaScript", 422));
-                }
-            }
-        }
-
-        async intercept(context: ExecutionContext, next: CallHandler<unknown>): Promise<Observable<unknown>> {
-            const fileInterceptor = await this.fileInterceptor.intercept(context, next);
-
-            const ctx = context.switchToHttp();
-            const file = ctx.getRequest().file;
-
-            const err = await this.validateFileContent(file);
-            if (err) {
-                return err;
-            }
-
-            return fileInterceptor;
+        intercept(...args: Parameters<NestInterceptor["intercept"]>): Observable<unknown> | Promise<Observable<unknown>> {
+            return this.fileInterceptor.intercept(...args);
         }
     }
     return mixin(Interceptor);
