+++ packages/api/cms-api/src/dam/files/file-validation-interceptor.service.ts
@@ -1,0 +1,36 @@
+import { CallHandler, ExecutionContext, HttpException, Injectable, NestInterceptor } from "@nestjs/common";
+import fs from "fs";
+import { Observable, throwError } from "rxjs";
+import * as util from "util";
+
+import { svgContainsJavaScript } from "./files.utils";
+
+const readFile = util.promisify(fs.readFile);
+const unlinkFile = util.promisify(fs.unlink);
+
+@Injectable()
+export class FileValidationInterceptor implements NestInterceptor {
+    async intercept(context: ExecutionContext, next: CallHandler<unknown>): Promise<Observable<unknown>> {
+        const ctx = context.switchToHttp();
+        const file = ctx.getRequest().file;
+
+        if (file && file.mimetype === "image/svg+xml") {
+            const fileContent = await readFile(file.path, { encoding: "utf-8" });
+
+            if (svgContainsJavaScript(fileContent)) {
+                // https://github.com/expressjs/multer/blob/master/storage/disk.js#L54-L62
+                const path = file.path;
+
+                delete file.destination;
+                delete file.filename;
+                delete file.path;
+
+                await unlinkFile(path);
+
+                return throwError(() => new HttpException("Rejected File Upload: SVG must not contain JavaScript", 422));
+            }
+        }
+
+        return next.handle();
+    }
+}
